New split module
